feat: merge iOS and tvOS AVFoundationPlayback tests
